requirements.txt
Django
channels
channels_redis
daphne
gunicorn
psycopg2-binary
whitenoise
python-dotenv
Pillow
celery[redis]
redis
dj-database-url
python-magic
drf-spectacular
django-cors-headers
django-filter
drf-nested-routers

